pyproject.toml
[project]
name = "grades-up"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beepy>=1.0.7",
    "colorlog>=6.9.0",
    "playwright>=1.52.0",
    "plyer>=2.1.0",
    "python-dotenv>=1.1.0",
    "twilio>=9.6.2",
]
